s21-day06/wip -> ex01 : probably done (minor changes (add fs module))
