+++ day06/ex01/internal/handlers/handle_admin.go
@@ -3,10 +3,9 @@ package handlers
 import (
 	"day06/ex01/internal/credentials"
 	"day06/ex01/internal/database"
-	"fmt"
+	"day06/ex01/internal/fs"
 	"log"
 	"net/http"
-	"os"
 	"strings"
 )
 
@@ -21,8 +20,8 @@ func HandleAdmin(w http.ResponseWriter, r *http.Request) {
 		articleName := r.FormValue("articleName")
 		articleLink := r.FormValue("articleLink")
 		if articleName != "" && articleLink != "" && strings.Contains(articleLink, ".md") {
-			if isCorrectArticle(articleLink) {
-				createArticle(articleLink)
+			if err := fs.CreateArticle(articleLink); err != nil {
+				log.Println(err)
 			}
 			if err := database.InsertArticle(articleName, articleLink); err != nil {
 				log.Println(err)
@@ -41,30 +40,4 @@ func validateAdmin(login, password string) bool {
 		return true
 	}
 	return false
-}
-
-func createArticle(name string) {
-	mdFile := fmt.Sprintf("./ui/md/%s", name)
-	if _, err := os.Create(mdFile); err != nil {
-		log.Println(err)
-	}
-
-	htmlName := strings.Split(name, ".md")[0] + ".html"
-	htmlFile := fmt.Sprintf("./ui/html/%s", htmlName)
-	if _, err := os.Create(htmlFile); err != nil {
-		log.Println(err)
-	}
-}
-
-func isCorrectArticle(link string) bool {
-	articles, err := database.GetArticles()
-	if err != nil {
-		log.Println(err)
-	}
-	for current, _ := range articles {
-		if current == link {
-			return false
-		}
-	}
-	return true
 }
